requirements.txt
fastapi
uvicorn[standard]
langchain-groq
langchain-huggingface
langchain-chroma
langchain-openai
pydantic
python-dotenv
PyMuPDF
sentence-transformers
sqlite3